flask_student_api/requirements.txt
Flask==2.3.2
flasgger==0.9.7.1
python-dotenv==1.0.0
Flask-Talisman==1.1.0
psycopg==3.1.18
pandas==2.2.2
openpyxl==3.1.2